requirements.txt
google-generativeai
yfinance
pandas
matplotlib
python-dotenv